+++ packages/requisitions/src/ResponseRequisition/DetailView/ContentArea.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { DataTable, NothingHere } from '@openmsupply-client/common';
-import { useResponseLines, ResponseLineFragment } from '../api';
+import { useResponse, ResponseLineFragment } from '../api';
 
 interface ContentAreaProps {
   onRowClick: null | ((line: ResponseLineFragment) => void);
 }
 
 export const ContentArea = ({ onRowClick }: ContentAreaProps) => {
-  const { columns, lines } = useResponseLines();
+  const { columns, lines } = useResponse.line.list();
 
   return (
     <DataTable
