+++ packages/requisitions/src/ResponseRequisition/api/hooks/document/index.ts
@@ -1,0 +1,13 @@
+import { useResponse } from './useResponse';
+import { useResponseFields } from './useResponseFields';
+import { useResponses } from './useResponses';
+import { useResponsesAll } from './useResponsesAll';
+import { useUpdateResponse } from './useUpdateResponse';
+
+export const Document = {
+  useResponse,
+  useResponseFields,
+  useResponses,
+  useResponsesAll,
+  useUpdateResponse,
+};
