+++ packages/requisitions/src/ResponseRequisition/api/hooks/document/useUpdateResponse.ts
@@ -1,0 +1,10 @@
+import { useQueryClient, useMutation } from '@openmsupply-client/common';
+import { useResponseApi } from '../utils/useResponseApi';
+
+export const useUpdateResponse = () => {
+  const queryClient = useQueryClient();
+  const api = useResponseApi();
+  return useMutation(api.update, {
+    onSuccess: () => queryClient.invalidateQueries(api.keys.base()),
+  });
+};
