+++ packages/requisitions/src/ResponseRequisition/api/hooks/line/index.ts
@@ -1,0 +1,9 @@
+import { useDeleteResponseLines } from './useDeleteResponseLines';
+import { useResponseLines } from './useResponseLines';
+import { useSaveResponseLines } from './useSaveResponseLines';
+
+export const Lines = {
+  useDeleteResponseLines,
+  useResponseLines,
+  useSaveResponseLines,
+};
