+++ packages/requisitions/src/ResponseRequisition/api/hooks/utils/index.ts
@@ -1,0 +1,11 @@
+import { useCreateOutboundFromResponse } from './useCreateOutboundFromResponse';
+import { useIsResponseDisabled } from './useIsResponseDisabled';
+import { useResponseApi } from './useResponseApi';
+import { useSupplyRequestedQuantity } from './useSupplyRequestedQuantity';
+
+export const Utils = {
+  useCreateOutboundFromResponse,
+  useIsResponseDisabled,
+  useResponseApi,
+  useSupplyRequestedQuantity,
+};
